requirements.txt
debugpy
gfpgan==1.3.8
gradio==3.44.3
insightface==0.7.3
numpy==1.24.3
onnx==1.14.1
onnxruntime-gpu
opencv-python==4.8.0.76
opennsfw2==0.10.2
pillow==10.0.1
protobuf==4.24.2
psutil==5.9.5
realesrgan==0.3.0
tensorflow
tqdm==4.66.1
ffmpeg-python